pyproject.toml
[project]
name = "portia-ai-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Ethan Weygang",email = "ethanweygang@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "portia-sdk-python (>=0.1.14,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
